requirements.txt
faiss-cpu==1.10.0
numpy==2.2.2
pandas==2.2.3
openai==1.61.1
psycopg2-binary==2.9.10
pgxnclient==1.3.2
psycopg2==2.9.10
python-dotenv==1.0.1
tabulate